# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Yeshshree Operations API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/approvals/delegations": {
      "get": {
        "tags": [
          "approvals"
        ],
        "summary": "List Delegations",
        "operationId": "list_delegations_api_v1_approvals_delegations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DelegationRead"
                  },
                  "type": "array",
                  "title": "Response List Delegations Api V1 Approvals Delegations Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "approvals"
        ],
        "summary": "Create Delegation",
        "operationId": "create_delegation_api_v1_approvals_delegations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DelegationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DelegationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/approvals/delegations/{delegation_id}": {
      "patch": {
        "tags": [
          "approvals"
        ],
        "summary": "Deactivate Delegation",
        "description": "DELETE-as-deactivate: only {is_active: false} is accepted (never-do: no DELETE).",
        "operationId": "deactivate_delegation_api_v1_approvals_delegations__delegation_id__patch",
        "parameters": [
          {
            "name": "delegation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Delegation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DelegationUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DelegationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/approvals": {
      "get": {
        "tags": [
          "approvals"
        ],
        "summary": "List Approvals",
        "operationId": "list_approvals_api_v1_approvals_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 200,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApprovalRead"
                  },
                  "title": "Response List Approvals Api V1 Approvals Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/approvals/{approval_id}/decide": {
      "post": {
        "tags": [
          "approvals"
        ],
        "summary": "Decide",
        "operationId": "decide_api_v1_approvals__approval_id__decide_post",
        "parameters": [
          {
            "name": "approval_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Approval Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecideRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/approvals/{approval_id}/override": {
      "post": {
        "tags": [
          "approvals"
        ],
        "summary": "Override",
        "description": "\u00a711.2 emergency override \u2014 post-facto override_review is auto-created.",
        "operationId": "override_api_v1_approvals__approval_id__override_post",
        "parameters": [
          {
            "name": "approval_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Approval Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OverrideRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/dashboards/achievement": {
      "get": {
        "tags": [
          "dashboards"
        ],
        "summary": "Achievement",
        "description": "Plan vs confirmed per line for a day (default: today).",
        "operationId": "achievement_api_v1_dashboards_achievement_get",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AchievementOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dashboards/yield": {
      "get": {
        "tags": [
          "dashboards"
        ],
        "summary": "Yield Dashboard",
        "description": "Good vs rejected by line/shift + reject reasons + imputed scrap value.",
        "operationId": "yield_dashboard_api_v1_dashboards_yield_get",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/YieldOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dashboards/sales": {
      "get": {
        "tags": [
          "dashboards"
        ],
        "summary": "Sales",
        "description": "Billed today / MTD from confirmed invoices (default month: current).",
        "operationId": "sales_api_v1_dashboards_sales_get",
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^\\d{4}-\\d{2}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dashboards/overview": {
      "get": {
        "tags": [
          "dashboards"
        ],
        "summary": "Overview",
        "description": "Management one-screen KPIs (default: today).",
        "operationId": "overview_api_v1_dashboards_overview_get",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AchievementLine": {
        "properties": {
          "line_id": {
            "type": "integer",
            "title": "Line Id"
          },
          "line_name": {
            "type": "string",
            "title": "Line Name"
          },
          "planned": {
            "type": "string",
            "title": "Planned"
          },
          "confirmed_good": {
            "type": "string",
            "title": "Confirmed Good"
          },
          "confirmed_rejected": {
            "type": "string",
            "title": "Confirmed Rejected"
          },
          "pct": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pct"
          },
          "shifts": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ShiftSplit"
            },
            "type": "object",
            "title": "Shifts"
          },
          "downtime_total_min": {
            "type": "integer",
            "title": "Downtime Total Min"
          },
          "top_downtime_reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TopDowntimeReason"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "line_id",
          "line_name",
          "planned",
          "confirmed_good",
          "confirmed_rejected",
          "pct",
          "shifts",
          "downtime_total_min",
          "top_downtime_reason"
        ],
        "title": "AchievementLine"
      },
      "AchievementOut": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/AchievementLine"
            },
            "type": "array",
            "title": "Lines"
          },
          "totals": {
            "$ref": "#/components/schemas/AchievementTotals"
          }
        },
        "type": "object",
        "required": [
          "date",
          "lines",
          "totals"
        ],
        "title": "AchievementOut"
      },
      "AchievementTotals": {
        "properties": {
          "planned": {
            "type": "string",
            "title": "Planned"
          },
          "confirmed_good": {
            "type": "string",
            "title": "Confirmed Good"
          },
          "confirmed_rejected": {
            "type": "string",
            "title": "Confirmed Rejected"
          },
          "pct": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pct"
          }
        },
        "type": "object",
        "required": [
          "planned",
          "confirmed_good",
          "confirmed_rejected",
          "pct"
        ],
        "title": "AchievementTotals"
      },
...
      "ApprovalRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "approval_type": {
            "type": "string",
            "title": "Approval Type"
          },
          "ref_type": {
            "type": "string",
            "title": "Ref Type"
          },
          "ref_id": {
            "type": "integer",
            "title": "Ref Id"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "required_roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Roles"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_by": {
            "type": "integer",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "decided_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "overridden_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overridden By"
          },
          "override_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Override Reason"
          }
        },
        "type": "object",
        "required": [
          "id",
          "approval_type",
          "ref_type",
          "ref_id",
          "payload",
          "required_roles",
          "status",
          "created_by",
          "created_at",
          "decided_at",
          "overridden_by",
          "override_reason"
        ],
        "title": "ApprovalRead"
      },
...
      "BilledBucket": {
        "properties": {
          "pcs": {
            "type": "string",
            "title": "Pcs"
          },
          "value": {
            "type": "string",
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "pcs",
          "value"
        ],
        "title": "BilledBucket"
      },
...
      "DecideRequest": {
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "approve",
              "decline",
              "hold"
            ],
            "title": "Decision"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "DecideRequest"
      },
...
      "DelegationCreate": {
        "properties": {
          "principal_user_id": {
            "type": "integer",
            "title": "Principal User Id"
          },
          "delegate_user_id": {
            "type": "integer",
            "title": "Delegate User Id"
          },
          "approval_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Type"
          },
          "valid_from": {
            "type": "string",
            "format": "date",
            "title": "Valid From"
          },
          "valid_to": {
            "type": "string",
            "format": "date",
            "title": "Valid To"
          }
        },
        "type": "object",
        "required": [
          "principal_user_id",
          "delegate_user_id",
          "valid_from",
          "valid_to"
        ],
        "title": "DelegationCreate"
      },
      "DelegationRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "principal_user_id": {
            "type": "integer",
            "title": "Principal User Id"
          },
          "delegate_user_id": {
            "type": "integer",
            "title": "Delegate User Id"
          },
          "approval_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Type"
          },
          "valid_from": {
            "type": "string",
            "format": "date",
            "title": "Valid From"
          },
          "valid_to": {
            "type": "string",
            "format": "date",
            "title": "Valid To"
          },
          "created_by": {
            "type": "integer",
            "title": "Created By"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "principal_user_id",
          "delegate_user_id",
          "approval_type",
          "valid_from",
          "valid_to",
          "created_by",
          "is_active"
        ],
        "title": "DelegationRead"
      },
      "DelegationUpdate": {
        "properties": {
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "is_active"
        ],
        "title": "DelegationUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OverrideRequest": {
        "properties": {
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "OverrideRequest"
      },
      "OverviewOut": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "achievement_pct": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Achievement Pct"
          },
          "billed_today_value": {
            "type": "string",
            "title": "Billed Today Value"
          },
          "yield_pct": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Yield Pct"
          },
          "open_anomalies": {
            "$ref": "#/components/schemas/OpenAnomalies"
          },
          "approvals_pending": {
            "type": "integer",
            "title": "Approvals Pending"
          },
          "open_override_reviews": {
            "type": "integer",
            "title": "Open Override Reviews"
          },
          "unmatched_gate_entries": {
            "type": "integer",
            "title": "Unmatched Gate Entries"
          },
          "outbox_backlog": {
            "$ref": "#/components/schemas/OutboxBacklog"
          },
          "holds_open": {
            "type": "integer",
            "title": "Holds Open"
          }
        },
        "type": "object",
        "required": [
          "date",
          "achievement_pct",
          "billed_today_value",
          "yield_pct",
          "open_anomalies",
          "approvals_pending",
          "open_override_reviews",
          "unmatched_gate_entries",
          "outbox_backlog",
          "holds_open"
        ],
        "title": "OverviewOut"
      },
...
      "SalesOut": {
        "properties": {
          "month": {
            "type": "string",
            "title": "Month"
          },
          "billed_today": {
            "$ref": "#/components/schemas/BilledBucket"
          },
          "billed_mtd": {
            "$ref": "#/components/schemas/BilledBucket"
          },
          "materials": {
            "items": {
              "$ref": "#/components/schemas/SalesMaterialRow"
            },
            "type": "array",
            "title": "Materials"
          },
          "awaiting_invoice": {
            "type": "integer",
            "title": "Awaiting Invoice"
          }
        },
        "type": "object",
        "required": [
          "month",
          "billed_today",
          "billed_mtd",
          "materials",
          "awaiting_invoice"
        ],
        "title": "SalesOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "YieldOut": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/YieldRow"
            },
            "type": "array",
            "title": "Rows"
          },
          "reject_reasons": {
            "items": {
              "$ref": "#/components/schemas/RejectReason"
            },
            "type": "array",
            "title": "Reject Reasons"
          },
          "imputed_scrap_value": {
            "type": "string",
            "title": "Imputed Scrap Value"
          },
          "overall": {
            "$ref": "#/components/schemas/YieldOverall"
          }
        },
        "type": "object",
        "required": [
          "date",
          "rows",
          "reject_reasons",
          "imputed_scrap_value",
          "overall"
        ],
        "title": "YieldOut"
      },
...
      "YieldRow": {
        "properties": {
          "line_id": {
            "type": "integer",
            "title": "Line Id"
          },
          "line_name": {
            "type": "string",
            "title": "Line Name"
          },
          "shift": {
            "type": "string",
            "title": "Shift"
          },
          "good": {
            "type": "string",
            "title": "Good"
          },
          "rejected": {
            "type": "string",
            "title": "Rejected"
          },
          "yield_pct": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Yield Pct"
          }
        },
        "type": "object",
        "required": [
          "line_id",
          "line_name",
          "shift",
          "good",
          "rejected",
          "yield_pct"
        ],
        "title": "YieldRow"
      },
...
    }
  }
}